refactored tests to reduce helper function duplication
